# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Constancia Backend API",
    "description": "Backend API contract for the Constancia frontend and Discord bot.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/campaigns/": {
      "get": {
        "operationId": "listCampaigns",
        "summary": "List campaigns",
        "tags": ["campaigns"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "discordGuildId": {
                            "type": "string"
                          },
                          "gameSystemId": {
                            "type": "string"
                          }
                        },
                        "required": ["id", "name", "discordGuildId", "gameSystemId"]
                      }
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCampaign",
        "summary": "Create a campaign",
        "tags": ["campaigns"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "discordGuildId": {
                    "type": "string"
                  },
                  "gameSystemId": {
                    "type": "string"
                  }
                },
                "required": ["name", "discordGuildId", "gameSystemId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordGuildId": {
                          "type": "string"
                        },
                        "gameSystemId": {
                          "type": "string"
                        }
                      },
                      "required": ["id", "name", "discordGuildId", "gameSystemId"]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{id}": {
      "get": {
        "operationId": "getCampaign",
        "summary": "Get campaign details",
        "tags": ["campaigns"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordGuildId": {
                          "type": "string"
                        },
                        "gameSystemId": {
                          "type": "string"
                        }
                      },
                      "required": ["id", "name", "discordGuildId", "gameSystemId"]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateCampaign",
        "summary": "Update a campaign",
        "tags": ["campaigns"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "gameSystemId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "discordGuildId": {
                          "type": "string"
                        },
                        "gameSystemId": {
                          "type": "string"
                        }
                      },
                      "required": ["id", "name", "discordGuildId", "gameSystemId"]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{id}/npcs/": {
      "get": {
        "operationId": "listNpcs",
        "summary": "List campaign NPCs",
        "tags": ["npcs"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "systemBlocks": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "additionalProperties": false,
                              "properties": {
                                "systemId": {
                                  "type": "string"
                                },
                                "blockType": {
                                  "type": "string"
                                },
                                "label": {
                                  "type": "string"
                                },
                                "value": {}
                              },
                              "required": ["blockType", "label", "value"]
                            }
                          },
                          "campaignId": {
                            "type": "string"
                          },
                          "facts": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "additionalProperties": false,
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "content": {
                                  "type": "string"
                                },
                                "sortOrder": {
                                  "type": "integer"
                                },
                                "npcId": {
                                  "type": "string"
                                },
                                "knownTo": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "additionalProperties": false,
                                    "properties": {
                                      "characterId": {
                                        "type": "string"
                                      },
                                      "discordUserId": {
                                        "type": "string"
                                      },
                                      "displayName": {
                                        "type": "string"
                                      },
                                      "secondaryLabel": {
                                        "type": "string"
                                      }
                                    },
                                    "required": [
                                      "characterId",
                                      "discordUserId",
                                      "displayName",
                                      "secondaryLabel"
                                    ]
                                  }
                                }
                              },
                              "required": ["id", "content", "sortOrder", "npcId", "knownTo"]
                            }
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "systemBlocks",
                          "campaignId",
                          "facts"
                        ]
                      }
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createNpc",
        "summary": "Create an NPC",
        "tags": ["npcs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "systemBlocks": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "systemId": {
                          "type": "string"
                        },
                        "blockType": {
                          "type": "string"
                        },
                        "label": {
                          "type": "string"
                        },
                        "value": {}
                      },
                      "required": ["blockType", "label", "value"]
                    }
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "facts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "content": {
                          "type": "string"
                        },
                        "sortOrder": {
                          "type": "integer"
                        }
                      },
                      "required": ["content"]
                    }
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "systemBlocks": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "properties": {
                              "systemId": {
                                "type": "string"
                              },
                              "blockType": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "value": {}
                            },
                            "required": ["blockType", "label", "value"]
                          }
                        },
                        "campaignId": {
                          "type": "string"
                        },
                        "facts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "content": {
                                "type": "string"
                              },
                              "sortOrder": {
                                "type": "integer"
                              },
                              "npcId": {
                                "type": "string"
                              }
                            },
                            "required": ["id", "content", "sortOrder", "npcId"]
                          }
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "description",
                        "systemBlocks",
                        "campaignId",
                        "facts"
                      ]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{id}/npcs/{npcId}/facts": {
      "post": {
        "operationId": "createNpcFact",
        "summary": "Add an NPC fact",
        "tags": ["npcs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  }
                },
                "required": ["content"]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "npcId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "sortOrder": {
                          "type": "integer"
                        },
                        "npcId": {
                          "type": "string"
                        }
                      },
                      "required": ["id", "content", "sortOrder", "npcId"]
                    }
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}